Moved graphql part to separate helper class
